Remove html_game_loop.rs
